requestDTO responseDTO 생성 및 엔티티 수정

+++ webBizMateProject/src/main/java/com/bizmate/project/domain/ProjectBoard.java
@@ -22,7 +22,7 @@ public class ProjectBoard extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE ,generator = "project_board_generator")
-    private Integer projectBoardId;
+    private Long projectBoardId;
 
     @Column(nullable = false)
     private String boardTitle;
